pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "rlog-generator"
version = "0.3.0"
description = "Generator of random logs for multiple types of technologies"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Apache-2.0"}
authors = [
    {name = "Fedele Mantuano", email = "mantuano.fedele@gmail.com"},
]
maintainers = [
    {name = "Fedele Mantuano", email = "mantuano.fedele@gmail.com"},
]
keywords = ["log", "generator", "random"]
classifiers = [
    "Intended Audience :: Developers",
    "Natural Language :: English",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]
dependencies = [
    "click>=8.1.7",
    "tqdm>=4.66.1",
    "pyyaml>=6.0.1",
    "faker>=37.0.0",
]

[project.urls]
Homepage = "https://github.com/WuerthPhoenix/log-generator"
Issues = "https://github.com/WuerthPhoenix/log-generator/issues"

[project.scripts]
rlog-generator = "rlog_generator.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["rlog_generator"]

[tool.black]
line-length = 88
target-version = ["py312"]

[tool.isort]
profile = "black"
line_length = 88

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true

#[tool.uv]
#python-version = "3.12"

[tool.pytest]

[tool.ruff]
line-length = 88

